+++ jest.config.js
@@ -3,5 +3,6 @@ module.exports = {
   setupFiles: ['./jest.setup.js'],
   snapshotSerializers: ['enzyme-to-json/serializer'],
   testRegex: '(/__tests__/.*(\\.|/)(test|spec))\\.jsx?$',
-  testPathIgnorePatterns: ['/node_modules/', '/lib/', '/docs/']
+  testPathIgnorePatterns: ['/node_modules/', '/lib/', '/docs/'],
+  //modulePathIgnorePatterns: ['/lib/']
 };