+++ packages/inline-dropdown/configure/src/defaults.js
@@ -70,6 +70,10 @@ export default {
     lockChoiceOrder: {
       settings: true,
       label: 'Lock Choice Order'
-    }
+    },
+    partialScoring: {
+      settings: true,
+      label: 'Partial Scoring'
+    },
   }
 };