+++ packages/explicit-constructed-response/configure/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/explicit-constructed-response-configure@1.2.0...@pie-element/explicit-constructed-response-configure@1.2.1) (2019-06-23)
+
+**Note:** Version bump only for package @pie-element/explicit-constructed-response-configure
+
+
+
+
+
+# [1.2.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/explicit-constructed-response-configure@1.1.3...@pie-element/explicit-constructed-response-configure@1.2.0) (2019-06-21)
+
+
+### Features
+
+* **general:** added choice position select to ditb, fixed ecr markup processing ([9376dac](https://github.com/pie-framework/pie-elements/commit/9376dac))
+* **general:** updated all the packages that have the config-ui as a dependency ([44061de](https://github.com/pie-framework/pie-elements/commit/44061de))
+
+
+
+
+
 ## [1.1.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/explicit-constructed-response-configure@1.1.2...@pie-element/explicit-constructed-response-configure@1.1.3) (2019-06-17)
 
 **Note:** Version bump only for package @pie-element/explicit-constructed-response-configure