+++ packages/match-list/controller/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-list-controller@1.1.4...@pie-element/match-list-controller@1.1.5) (2019-06-20)
+
+**Note:** Version bump only for package @pie-element/match-list-controller
+
+
+
+
+
+## [1.1.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-list-controller@1.1.3...@pie-element/match-list-controller@1.1.4) (2019-06-19)
+
+**Note:** Version bump only for package @pie-element/match-list-controller
+
+
+
+
+
 ## [1.1.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/match-list-controller@1.1.2...@pie-element/match-list-controller@1.1.3) (2019-06-18)