+++ packages/pie-models/CHANGELOG.md
@@ -3,6 +3,79 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.3.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@5.3.1...@pie-element/pie-models@5.3.2) (2019-06-24)
+
+**Note:** Version bump only for package @pie-element/pie-models
+
+
+
+
+
+## [5.3.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@5.3.0...@pie-element/pie-models@5.3.1) (2019-06-24)
+
+**Note:** Version bump only for package @pie-element/pie-models
+
+
+
+
+
+# [5.3.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@5.2.4...@pie-element/pie-models@5.3.0) (2019-06-23)
+
+
+### Features
+
+* update documentation for number-line ([f7f1be8](https://github.com/pie-framework/pie-elements/commit/f7f1be8))
+
+
+
+
+
+## [5.2.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@5.2.3...@pie-element/pie-models@5.2.4) (2019-06-21)
+
+**Note:** Version bump only for package @pie-element/pie-models
+
+
+
+
+
+## [5.2.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@5.2.2...@pie-element/pie-models@5.2.3) (2019-06-20)
+
+**Note:** Version bump only for package @pie-element/pie-models
+
+
+
+
+
+## [5.2.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@5.2.1...@pie-element/pie-models@5.2.2) (2019-06-20)
+
+**Note:** Version bump only for package @pie-element/pie-models
+
+
+
+
+
+## [5.2.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@5.2.0...@pie-element/pie-models@5.2.1) (2019-06-19)
+
+
+### Bug Fixes
+
+* **scoring-and-feedback:** fixed scoring and added alternate responses ([dd0bb02](https://github.com/pie-framework/pie-elements/commit/dd0bb02))
+
+
+
+
+
+# [5.2.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@5.1.4...@pie-element/pie-models@5.2.0) (2019-06-19)
+
+
+### Features
+
+* **extras:** added row labels support and alternate responses ([280028f](https://github.com/pie-framework/pie-elements/commit/280028f))
+
+
+
+
+
 ## [5.1.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@5.1.3...@pie-element/pie-models@5.1.4) (2019-06-17)
 
 **Note:** Version bump only for package @pie-element/pie-models